// src/types/inventoryVoucher/GstBillTypes.d.ts
interface ItemData {
  item_name: string;
  item_qnty: string;
  item_unit: string;
  item_hsn: string;
  item_rate: string;
  item_tot: string;
}
// ...
export interface InvoiceModalProps {
  showInvoice: boolean;
  setShowInvoice: Dispatch<SetStateAction<boolean>>;
  invoiceData: InvoiceData | null;
  setInvoiceData: Dispatch<SetStateAction<InvoiceData | null>>;
}

export interface InvoiceData {
  Sales_Id: 6;
  Sales_Date: string;
  Sales_No: string;
  Party_Name: string;
  Gross_Amt: number;
  Cgst_Rate: number;
  Cgst_Amt: number;
  Sgst_Rate: number;
  Sgst_Amt: number;
  Igst_Amt: number;
  Round_Amt: number;
  Discount: number;
  ItemData: {
    Item_Name: string;
    Item_Qnty: number;
    Item_Unit: string;
    Item_Rate: number;
    Item_Hsn: number;
    Item_Tot: number;
  }[];
}